Create kali_anti_virus.sh
This is a custom antivirus script my partner and I wrote for our FullStack / CSU Cyber Security Boot Camp final project.  This is the Kali version that we tested with our Kali VM. The functionality of this script is to take a target hash as an argument, search the user's home directory (and all subdirectories recursively) for the malicious file. If any files have matching hashes, they are stripped of their permissions and moved to the Quarantine directory. This script will work for any number of instances of the malicious file throughout the user's files.